Refactor script to accept config as param and support updatedSince attribute
Update the script to accept two new parameters: -c configFilePath and
-u updatedSince. The first allow to provide a script from a different
path than the one where the script is located. The second allow to use
the AppNexus parameter updated_since in the format of YYYY_MM_DD_HH to
filter how far in time we go to retrieve log files. Also improved
logging of configuration and added support for s3 parameters in prep
for supporting s3 direct upload.
Added a config file sample