Add support for s3 paths as dataDir. If dataDir is give as s3://bucket-name/path-to/an-data we will store the feeds into this path directly on s3 instead of locally. All duplicate and checksum tests continues to work as before. We are doing a best-attempt to detect checksum of remote s3 files before we decide to upload them (based on self stored anchecksum metadata value, as well a falling back to the ETag value that in most cases will be the md5 checksum of the file.
Local space requirements is limited to the size of a single file, one at a time. We could have stream the file directly to S3 without a local file, but then the checksum process would be harder to do, because we can't tell S3 if the whole file was uploaded or not without downloading it again and computing the checksum locally, which would be redundant. So instead we just download to a temp file, upload to s3, and then delete the temp file.

Some new features added on the way:
- -s --split to split data into daily folders (might not be needed because of the way that s3 files are marked, but in case it is needed we can turn this on)
- make default path of config file to be local working dir, instead of next to the script itself.
- improved logging
- added a requirements.txt file (generated using pigar) to ensure it is easy to install all dependencies required to run the script.